+++ Encoder.h
@@ -25,9 +25,9 @@
 void encoderInit();
 
 // Gets the total distance traveled for this particular encoder
-void getEncoderDistances(uint64_t * encoderDistances);
+void getEncoderDistances(int64_t * encoderDistances);
 
 // Gets the average number of ticks for the pair of encoders (used for PID of speed)
-void getAverageEncoderTicks(int32_t * avEncoderTicks);
+void getAverageEncoderTicks(int64_t * avEncoderTicks);
 
 #endif//_ENCODER_H_
